requirements.txt
pyspark>=3.4.0
delta-spark>=3.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0
mlflow>=2.8.0
shap>=0.43.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
